:recycle: Rename Module To Avoid Shadowing
