+++ wsic/__init__.py
@@ -4,7 +4,7 @@ __author__ = """John Pocock"""
 __email__ = "j.c.pocock@warwick.ac.uk"
 __version__ = "0.6.1"
 
-from . import codecs, magic, metadata, multiprocessing, readers, types, writers
+from . import codecs, magic, metadata, multiproc, readers, types, writers
 
 __all__ = [
     "codecs",
@@ -13,5 +13,5 @@ __all__ = [
     "types",
     "writers",
     "metadata",
-    "multiprocessing",
+    "multiproc",
 ]
