+++ wsic/readers.py
@@ -16,7 +16,7 @@ from wsic.codecs import register_codecs
 from wsic.enums import Codec, ColorSpace
 from wsic.magic import summon_file_types
 from wsic.metadata import ngff
-from wsic.multiprocessing import Queue
+from wsic.multiproc import Queue
 from wsic.types import PathLike
 from wsic.utils import (
     block_downsample_shape,
